requirements.txt
fastapi
uvicorn
python-multipart
uuid
mysql-connector-python
python-dotenv
jinja2
aiofiles